requirements.txt
asv==0.6.3
django_mongodb @ git+https://github.com/mongodb-labs/django-mongodb
pre-commit
